Use temporary files in tests to mock real files like the whitelist
Now we can verify that the files were changed exactly as expected.
I looked into other solutions, such as passing in a `&mut Vec<u8>` to
the test config, but it required some crazy concurrency stuff like
`RwLock<Cursor<&'static mut Vec<u8>>>`, which is annoying to create,
and the lifetime had to be static so that it can be shared among threads
(meaning that you would have to create a `static mut` variable, which is
inherently unsafe). So this solution is the cleanest, and also removes
the `Box<BufRead>`/`Box<Write>` indirection.

The `tempfile` dependency is only a dev-dependency, so it will not be
in the final binary.

Also added an `is_test()` function to the config.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>