Use `rust-embed` to embed the web interface files
Currently it uses the `master` branch of the repo, because the changes
have yet to be released in a new version. The changes were authored by
myself, so I believe they are stable. When the new version is released
this will be updated.

The web files are to be placed under the `web` folder.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>