Fix deadlock caused by concurrent FTL API accesses
FTL turns on a global thread lock when an API call runs, which
effectively makes all API calls run in sequence and blocks the resolver
until the call is finished. Because of this, opening one connection then
opening a second and reading from the second causes a deadlock because
FTL is trying to write to the first one while the API is waiting for the
second connection's data.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>