Expose Core's version and add TestConfigBuilder
TestConfigBuilder was created by extracting the functionality of
creating the data needed for `Config::Test` into a new builder. It can
be used to test smaller pieces of functionality without going through
a network request, such as testing `read_core_version`.

`web::WebAssets` was made public so the web version can be read off in
a future commit.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>