Improve CI caching
Caches now use Cargo.lock instead of Cargo.toml, and because the lock
file guarantees the cached dependencies are the same, caches can be used
across branches. The job name is still left in there to try and keep
the cache sizes down by only caching that target's build files.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>