Temporary fix for multiple temp file reads
When calling `try_clone`, the new `File` struct has the same
underlying handle to the file, and the same cursor. So after the
file is read the first time, any new reads will start at the place
the first read left off.

The more permanent fix is to use `NamedTempFile` and call `reopen`:
https://github.com/Stebalien/tempfile/issues/50#issuecomment-378996380

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>