Test an edge case when writing settings, and improve test file support
The edge case was fixed and a test added.

`TestFile` now has a `assert_expected` method which can be used to check
that the file contents match the expected contents. This makes it easy
to test IO outside of a full `TestBuilder` test.

Also added a new error, `InvalidSettingValue`, used when a setting fails
validation, such as when writing to the config file.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>