Implement `/stats/history` with shared memory and cursors
Work still needs to be done on the tests and documentation for the
filter functions.

The endpoint now supports pagination using cursors, and it works with
all of the other filters.

It builds an iterator based off of the filters, which it uses to
efficiently find all of the queries that should be returned.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>