Use the shared lock as a mutex and locally count read locks
FTL's lock changed to a robust mutex, so that it can handle if a process
holding the lock dies. Due to that, the API needs to manage read locks
internally. This does make the API of `FtlMemory` a little nicer.

The ShmLock is now a property of `FtlMemory::Production` and holds the
number of read locks currently held by the API. If a read lock is
requested and there are no read locks, it will lock the shared mutex and
then it will hand out the read lock. When the last read lock is dropped,
it will unlock the shared mutex.

It is planned for FTL to communicate with the API to notify it if it is
waiting for a write lock, so the API can avoid starving FTL.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>