Use a bool and condition variable to detect if FTL is waiting for a lock
FTL will set the boolean and start waiting for the lock. If the API
tries to make new internal read locks, it will check if FTL is waiting
and if so, wait upon FTL's condition variable before continuing with the
lock process. Once all of the API's locks are finished, FTL will get the
lock, set the boolean to false, and notify the condition variable so the
API can continue the locking process.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>