Fix locking issues with new lock handler thread
Lock handling is moved to a dedicated thread, since pthread does not
allow locks to be locked and unlocked in different threads. This also
removes the need for the condition variable, since lock requests that
defer to FTL are queued up for locking once FTL gets the lock.

Converting shmem errors has been moved into a impl of
`From<shmem::Error>` for `Error`, which cleans up the code a lot.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>