Add support for the FTL database in history
If a timespan is given to the history endpoint, it will use the database
if it runs out of queries in shared memory. If a timespan is not given,
the default timespan is the last 24 hours, which is completely stored in
shared memory, so the database does not need to be used.

The docker containers needed to have libsqlite3 to compile with sqlite
support, so they have been updated.

The test timestamps have been updated to offset the in-memory queries by
24 hours from the database queries.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>