Support BrowserRouter use in the web interface
BrowserRouter makes the pages of the web interface seem like real pages
(ex. `/admin/dashboard`), even if they don't exist on the backend. This
requires the backend to redirect such pages to `index.html` instead of
returning a 404. A 404 will still be returned if the missing file has
a file extension.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>