Add support for enabling/disabling blocking
In addition, there is support for disabling only for a fixed time
period. This is achieved by scheduling the re-enable work on another
thread.

The library used is a fork I just made which fixes a very small API
issue which would have prevented us from using it (it required a mutable
borrow when it didn't need one). It should hopefully get merged in as
soon as the maintainer sees the PR.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>