Add support for API_EXCLUDE_* SetupVars entries in the history endpoint
The `API_EXCLUDE_DOMAINS` and `API_EXCLUDE_CLIENTS` settings are used in
a few endpoints, even on the PHP web interface. However, they were not
used in the query log.

When getting excluded clients/domains, the data structure used was
changed from a Vec to HashSet for performance reasons.

Support for this filter in the database is coming in the next commits.

Closes #93

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>