Move database filters into separate files
The database filters now exist next to the non-database filters, such
as the domain filter living in domain.rs with the non-database domain
filter. The tests for the filters have also moved with them.

The list of blocked statuses was made static, so there is a single
reference to the statuses that are considered blocked.

`connect_to_test_db` was moved to the `databases::ftl` module.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>