Use armv7 for armhf binaries
Debian uses armv7 for the armhf architecture, and FTL's armhf build uses
armv7 (because it uses the armhf gcc in Debian). The API currently uses
Rust's arm-unknown-linux-gnueabihf target for the armhf build, but that
uses armv6 in order to be compatible with older Raspberry Pi's such as
the Pi 1B and Pi Zero. The armv7-unknown-linux-gnueabihf target is the
equivalent of Debian's armhf architecture. Because FTL uses armv7 for
the armhf build, API should also use armv7 for its armhf build.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>