Move database summary endpoint to /stats/database/summary
The database endpoints should be placed under a new path to
differentiate them from the in-memory (FTL) endpoints. The history
endpoint is an exception, because it uses a combination of in-memory and
database data to handle calls. Only database endpoints which do not rely
heavily on shared memory should use this new path.

This change fixes the route rating issue between the two summary
endpoints.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>