Update dependencies
Everything except nix was updated within the semver constraints (nix
updated from 0.12 to 0.13, which is technically a breaking change
according to how Rust uses semver).

This also has the effect of remaking the CI caches, which will help
armhf build times. The current armhf cache is for
arm-unknown-linux-gnueabihf, but the build was changed to
armv7-unknown-linux-gnueabihf. This caused the build time of armhf to be
longer than the other builds by about a minute. When the build for this
commit occurs, it will update the cache with the new architecture.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>