Center database overTime timestamps, and improve timestamp alignment
The timestamps supplied by FTL are centered in their overTime slots, so
the database endpoints should also align the timestamps.

The from/until alignment logic was simplified a bit. Now the `from`
timestamp will be floored to the interval, and the `until` timestamp
be raised to the interval (the range can only be expanded).

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>