Add a new reply_result method for simpler reply generation
Some endpoints (mostly the newer or refactored ones) will have the
actual endpoint method call another "impl" method to do the actual work.
The "impl" method will return the results in a nice form (not JSON).

This makes the endpoints easily testable by just calling the "impl"
method instead of dispatching a request. However, the request handler
will look something like this:

```rust
reply_data(my_endpoint_impl(input)?)
```

The "impl" function usually returns a `Result` because there may be
errors when doing the work, and the possible errors must be handled.
This is done by `?`, which "tries" to unwrap the data but returns the
error if it exists. See [here][try] for more information on the try
feature. When an `Error` is returned, it eventually gets run through
`reply_error`, but this indirection can be confusing. The indirection
can be simplified with a new reply method:

```rust
reply_result(my_endpoint_impl(input))
```

This new method correctly handles both cases of the `Result` without
causing any indirection.

[try]: https://doc.rust-lang.org/book/ch09-02-recoverable-errors-with-result.html#a-shortcut-for-propagating-errors-the--operator

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>