Statically link SQLite and fix MUSL linking
Until now, SQLite and its dependencies were dynamically linked, even in
MUSL builds which are supposed to be fully static. This caused even MUSL
builds to fail to run in CentOS, which has issues with the linking
requirements. By statically linking SQLite, the MUSL build is fully
static and can run on CentOS.

The installer will have to choose the MUSL build if running on CentOS,
or it could always choose MUSL if using x86_64.

Note that FTL also statically links SQLite, so we are one step closer
to removing an explicit SQLite dependency in Core (The shell scripts
still require it for interacting with the database).

This change increases the binary size by about 2MB.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>