Replace glibc x86_64 with MUSL
This change replaces the normal x86_64 build with the MUSL build. The
MUSL build works on all 64bit platforms, including CentOS, so there is
little point in having both normal (glibc) and MUSL x86_64 builds.

By making this change, Core does not have to change the binary that it
pulls in order to use the MUSL binary.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>