Add `Service` for dependency injection via Rocket request guards
`Service` is a way to do dependency injection (and therefore mocking) of
services via Rocket request guards. When testing, the request guard will
use the mock version if it is in Rocket's state and fall back to the
implementation version if the mock is not there. When used in
production, there is no overhead as the test code is conditionally
compiled.

This makes it possible to test the endpoints via `TestBuilder` without
calling real service code (such as making real database calls).

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>