Use a repository (service) for List database interactions
This abstracts the actual database interactions from the business logic
of the list service. The business logic can be tested separate from the
database layer.

The library `mock-it` is added to provide mocks. Its mocks are thread
safe, which is required for putting them in Rocket's state. Work will
be done in the library to improve its ergonomics, such as adding a
procedural macro for deriving the mock from a trait.

A future commit will put the list business logic behind its own service
and will restore some of the deleted tests.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>