Put the List code behind a service
The List code was moved into a new ListServiceImpl struct. List is now
just an enum, but it retains the `accepts` method.

The Service code had to be refactored a bit to handle some new
constraints, namely having lifetimes in Impl structs. That took a long
time to figure out, and the solution requires a nightly feature. Now all
Impl services need to have a lifetime parameter which represents the
lifetime of the request. If they don't really need it, they can store it
in PhantomData for no cost.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>