Fix crash when reading from missing config files
On startup we read from a few config files, including FTL's config. That
config file may not exist, which caused the API to crash as it couldn't
open the file.

Now the default value of the entry will be returned if the file was not
found. Other errors, such as permissions errors, will still be returned.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>